+++ lib/components/walk/slider_container.dart
