+++ lib/views/walk_view.dart
@@ -3,19 +3,19 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:provider/provider.dart';
 import 'package:yeogijeogi/components/common/custom_button.dart';
 import 'package:yeogijeogi/components/common/custom_scaffold.dart';
-import 'package:yeogijeogi/components/onboarding/slider_container.dart';
+import 'package:yeogijeogi/components/walk/slider_container.dart';
 import 'package:yeogijeogi/utils/palette.dart';
-import 'package:yeogijeogi/view_models/home_view_model.dart';
+import 'package:yeogijeogi/view_models/walk_view_model.dart';
 
-class HomeView extends StatelessWidget {
-  const HomeView({super.key});
+class WalkView extends StatelessWidget {
+  const WalkView({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final HomeViewModel homeViewModel = context.watch<HomeViewModel>();
+    final WalkViewModel walkViewModel = context.watch<WalkViewModel>();
 
     return CustomScaffold(
-      isLoading: homeViewModel.isLoading,
+      isLoading: walkViewModel.isLoading,
       body: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
@@ -57,18 +57,18 @@ class HomeView extends StatelessWidget {
           SliderContainer(
             title: "어떤 풍경을 찾아볼까요?",
             criteria: ["자연", "상관없음", "도시"],
-            value: homeViewModel.sceneryLevel,
+            value: walkViewModel.sceneryLevel,
             onChanged: (value) {
-              homeViewModel.updateSceneryLevel(value);
+              walkViewModel.updateSceneryLevel(value);
             },
           ),
           SizedBox(height: 24.h),
           SliderContainer(
             title: "산책 강도를 선택해 주세요.",
             criteria: ["가벼운", "상관없음", "운동되는"],
-            value: homeViewModel.walkingLevel,
+            value: walkViewModel.walkingLevel,
             onChanged: (value) {
-              homeViewModel.updateWalkingLevel(value);
+              walkViewModel.updateWalkingLevel(value);
             },
           ),
           Spacer(),
